requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.4
python-multipart>=0.0.9
requests>=2.31.0
apscheduler>=3.10.4
pydantic-settings>=2.2.1
aiofiles>=23.2.1
